+++ app/subscribers/api_prometheus_subscriber.rb
@@ -22,7 +22,7 @@ class ApiPrometheusSubscriber < ActiveSupport::Subscriber
   end
 
   def connection_failure(event)
-    exception = event.payload[:exception] if event.respond_to?(:exception)
+    exception = event.payload[:exception]
     message = exception.respond_to?(:message) ? exception.message : exception.to_s
 
     Prometheus::Client.registry
@@ -32,6 +32,14 @@ class ApiPrometheusSubscriber < ActiveSupport::Subscriber
     Prometheus::Client.registry
                       .get(:api_connection_failure)
                       .increment(labels: { message: })
+
+    Log.error(
+      "API connection failure: #{exception.message} - #{exception.class.name}",
+      {
+        request_status: 'error',
+        status: 503
+      }
+    )
   end
 
   def service_exception(event)
@@ -42,7 +50,15 @@ class ApiPrometheusSubscriber < ActiveSupport::Subscriber
 
     Prometheus::Client.registry
                       .get(:api_service_exception)
-                      .increment(labels: { status: })
+                      .increment(labels: { message: exception.to_s })
+
+    Log.error(
+      "API service exception: #{exception.message} - #{exception.class.name}",
+      {
+        request_status: 'error',
+        status: exception.respond_to?(:status) ? exception.status : 502
+      }
+    )
   end
 
   private